Mapper no Navio

+++ src/main/java/terrabuio/heitor/abysslogv2/internal/recurso/dto/request/RecursoRequest.java
@@ -7,7 +7,7 @@ public record RecursoRequest(
         @NotBlank(message = "O nome do recurso é obrigatório")
         String nome,
         @NotNull(message = "A quantidade é obrigatória")
-        Double quantidade,
+        Integer quantidade,
         @NotNull(message = "O ID da expedição é obrigatório")
         Long idExpedicao
 ) {}
