+++ src/main/java/terrabuio/heitor/abysslogv2/internal/recurso/dto/response/RecursoResponse.java
@@ -3,6 +3,6 @@ package terrabuio.heitor.abysslogv2.internal.recurso.dto.response;
 public record RecursoResponse(
         Long id,
         String nome,
-        Double quantidade,
+        Integer quantidade,
         Long idExpedicao
 ) {}
